probleme dans bdd mais service dispo

+++ Fournisseur/Infrastructure/src/main/java/DB/DataBase.java
@@ -12,6 +12,7 @@ import entities.Product;
 import entities.ProductSheet;
 import entities.Supplier;
 import infraInterface.CRUDI;
+import infraInterface.FactoryI;
 import infraInterface.SearchDBEntities;
 import factoryImpl.CreateInstanceBoutique;
 
@@ -40,21 +41,46 @@ public class DataBase implements CRUDI, SearchDBEntities {
 	
 	
 	
-	private void createTables() throws SQLException {
+	public void createTables() throws SQLException {
 		Statement stmt = conn.createStatement();
+		
+		//drop table before 
+		this.dropTables();
 		// create table
-	    stmt.executeUpdate("Create table PRODUCT (idProduct int IDENTITY(1,1) primary key, name varchar(50),idDescription int REFERENCES DESCRIPTION (idDescription), idSupplier int REFERENCES DESCRIPTION (idSupplier)");
-	    stmt.executeUpdate("Create table SUPPLIER (idSupplier int IDENTITY(1,1) primary key, name varchar(50) )");
-	    stmt.executeUpdate("Create table DESCRIPTION (idDescription int IDENTITY(1,1) primary key, DescritionText varchar(255) )");
+	    stmt.executeUpdate("Create table SUPPLIER (idSupplier INTEGER GENERATED ALWAYS AS IDENTITY (START WITH 1, INCREMENT BY 1) PRIMARY KEY, name varchar(50) )");
+	    stmt.executeUpdate("Create table DESCRIPTION (idDescription INTEGER GENERATED ALWAYS AS IDENTITY (START WITH 1, INCREMENT BY 1) PRIMARY KEY, DescriptionText varchar(255) )");
+	    stmt.executeUpdate("Create table PRODUCT (idProduct INTEGER  GENERATED ALWAYS AS IDENTITY (START WITH 1, INCREMENT BY 1) PRIMARY KEY, name varchar(50), price int, idDescription INTEGER REFERENCES DESCRIPTION (idDescription), idSupplier INTEGER REFERENCES SUPPLIER (idSupplier))");
 
+	    // exemple de remplissage de table
+	    FactoryI factory = new CreateInstanceBoutique();
+	    
+	    addProduct(factory.createProduct("a", "des trucs en description", 10));
+	    addProduct(factory.createProduct("b", "description de choses et d'autres", 1));
+	    addProduct(factory.createProduct("c", "produit excellent", 190));
+	    addProduct(factory.createProduct("d", "produit excellentissime", 1900));
 	}
 	
-	private void dropTables() throws SQLException{
+	public void dropTables() throws SQLException{
 		Statement stmt = conn.createStatement();
 		// drop table
-	    stmt.executeUpdate("Drop Table PRODUCT");
-	    stmt.executeUpdate("Drop Table SUPPLIER");
-	    stmt.executeUpdate("Drop Table DESCRIPTION");
+		ResultSet res;
+		
+		res = conn.getMetaData().getTables(null, "APP", "PRODUCT".toUpperCase(), null);
+		if(res.next()){
+			stmt.executeUpdate("Drop Table PRODUCT");
+		}
+	    
+		res = conn.getMetaData().getTables(null, "APP", "SUPPLIER".toUpperCase(), null);
+		if(res.next()){
+			stmt.executeUpdate("Drop Table SUPPLIER");
+		}
+		
+		res = conn.getMetaData().getTables(null, "APP", "DESCRIPTION".toUpperCase(), null);
+		if(res.next()){
+			stmt.executeUpdate("Drop Table DESCRIPTION");
+		}
+		
+	  
 	}
 	
 	public void addSupplier(Supplier s) throws SQLException{
@@ -65,14 +91,16 @@ public class DataBase implements CRUDI, SearchDBEntities {
 	public void addProduct(Product p) throws SQLException{
 		Statement stmt = conn.createStatement();
 		
-		ResultSet rs = stmt.executeQuery("SELECT idDescription FROM DESCRIPTION WHERE name ="+p.getDescription().getDescriptionText());
+		ResultSet rs = stmt.executeQuery("SELECT idDescription FROM DESCRIPTION WHERE DescriptionText ='"+p.getDescription().getDescriptionText().replaceAll("'", "")+"'");
 		
 		if(rs.next() != false){
 			int idDescription = rs.getInt(0);
-			stmt.executeUpdate("insert into PRODUCT values ("+p.getName()+","+idDescription+")");
+			stmt.executeUpdate("insert into PRODUCT values ("+p.getName()+","+p.getPrice()+","+idDescription+")");
 		}else{
-			
-			//stmt.executeUpdate("insert into PRODUCT values ("+p.getName()+","+,'tom')");
+			stmt.executeUpdate("insert into DESCRIPTION (DescriptionText) values('"+p.getDescription().getDescriptionText().replaceAll("'", "")+"')");
+			rs = stmt.executeQuery("SELECT idDescription FROM DESCRIPTION WHERE DescriptionText ='"+p.getDescription().getDescriptionText().replaceAll("'", "")+"'");
+			int idDescription = rs.getInt(0);
+			stmt.executeUpdate("insert into PRODUCT (name,price) values ("+p.getName()+","+p.getPrice()+","+idDescription+")");
 		}
 		/*stmt.executeUpdate("insert into PRODUCT values ("+p.getName()+","+,'tom')");
 	    stmt.executeUpdate("insert into users values (2,'peter')");*/
@@ -95,7 +123,7 @@ public class DataBase implements CRUDI, SearchDBEntities {
 		CreateInstanceBoutique factory = new CreateInstanceBoutique(); 
 		Statement stmt = conn.createStatement();
 		 ResultSet res = stmt.executeQuery("SELECT name,price FROM Product WHERE name ='"+name+"'");
-		 if(res.next() != false){
+		 if(res.next()){
 			return factory.createProduct(res.getObject(0).toString(),Float.parseFloat(res.getObject(1).toString())); 
 		 }
